+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 2026-07-18 22:11
+- Completing and Refactoring _Types/uCHXMatrix.pas_ and adding a little test program.
+- Adding ucWorleyNoise.pas, a class wich generate Worley Noise and test program.
+
+
 ## 2026-07-17 00:17
 ### uCHXMath.pas
 - Some physics constants, not sure if separate this kind of constants and